fixed x86-64 compilation
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@11183 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ util/filters/include/filters/realtime_vector_circular_buffer.h
@@ -83,7 +83,7 @@ public:
       return cb_.back();
   };
   
-  unsigned int size(){  return std::min(counter_, cb_.size());};
+  unsigned int size(){  return std::min(counter_, (unsigned int)cb_.size());};
   bool empty(){return cb_.empty();};
   T& at(size_t index){return cb_.at(index);};
   T& operator[](size_t index){return cb_[index];}
